Round dagsats to long

+++ src/main/kotlin/no/nav/helse/sykepenger/Dagsats.kt
@@ -1,6 +1,5 @@
 package no.nav.helse.sykepenger
 
-import java.math.BigDecimal
 import java.time.LocalDate
 
-data class Dagsats(val dato: LocalDate, val sats: BigDecimal, val skalUtbetales: Boolean)
+data class Dagsats(val dato: LocalDate, val sats: Long, val skalUtbetales: Boolean)
